step1 refactor: read/write service 분리, controller 분리
