+++ src/main/java/nextstep/subway/applicaion/dto/StationResponse.java
@@ -3,10 +3,10 @@ package nextstep.subway.applicaion.dto;
 import java.time.LocalDateTime;
 
 public class StationResponse {
-    private Long id;
-    private String name;
-    private LocalDateTime createdDate;
-    private LocalDateTime modifiedDate;
+    private final Long id;
+    private final String name;
+    private final LocalDateTime createdDate;
+    private final LocalDateTime modifiedDate;
 
     public StationResponse(
             Long id, String name, LocalDateTime createdDate, LocalDateTime modifiedDate) {
